+++ packages/router/src/types/index.ts
@@ -30,5 +30,8 @@ export {
   ResponseHandlerOptions,
   RemoveObserver,
   Navigation,
-  CurrentResponse
+  CurrentResponse,
+  Cancellable,
+  CancelActiveNavigation,
+  RemoveCancellable
 } from "./curi";